feat: Add support for web login using resty library
This commit adds the `resty` library to handle the web login process. It also updates the `WebLogin` function to use the `resty` client instead of the standard library's `http` package. The `WebLogin` function now makes use of the `LoginAuthResponse` and `LoginPswResponse` structs to parse the XML responses from the server.

The `AuthSms` and `TOTPAuth` functions have also been updated to use the `resty` client.

The commit also includes the implementation of the `ECAgentToken` function to handle the ECAgent token request.

Finally, the `createRestyClient` function has been added to create a new `resty` client with the base URL and TLS configuration.

This commit addresses issue #1234.